+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ##0.1.8 (-- --- 2013):
 - Problem with replacement and mixed datatype (https://github.com/PierreRambaud/GotCms/issues/55)
 - Use cms_version instead version (https://github.com/PierreRambaud/GotCms/issues/57)
+- Optimize the helper "documents" (https://github.com/PierreRambaud/GotCms/issues/58)
 
 ##0.1.7 (30 Apr 2013):
 - Get latest version can failed on certain php version (https://github.com/PierreRambaud/GotCms/issues/53)